GLMMng

Documentation for GLMMng.

GLMMng.GHnorm

GLMMng.pdeviance

GLMMng.tblrow

GLMMng.GHnorm — Method

GHnorm(k::Int)

Return a k-vector of abscissae and weights for normalized Gauss-Hermite quadrature

The function values are stored (memoized) when first evaluated. Subsequent evaluations for the same k have very low overhead.

GLMMng.pdeviance — Method

pdeviance(m::GLMMmod)

Return the penalized deviance.

The deviance is the StatsAPI.deviance value for the model without random effects (i.e. as a GLM model) and the penalty is the squared length of the vector of modes of the spherical random effects.

GLMMng.tblrow — Method

tblrow(::DistLink, y, η, offset=0)

Return a NamedTuple of μ, dev, rtwwt, wwres, wwresp from scalar y, η, and offset